spike: quick correction of series order
Series order is determined by `c:ser/c:order/@val` and not
`c:ser/c:idx/@val` as previously thought. The `idx` value is more like
a local 'id' than an index (offset) in a sequence as the term is
commonly used in Python.

I think this gets it done, but this should be reviewed carefully and
perhaps tested more thoroughly before merging.